refactor(modal): only one modal should be open at any time
https://github.com/graycoreio/daffodil/issues/1456
