+++ libs/design/modal/src/service/specs/modal.service.spec.ts
@@ -16,6 +16,7 @@ import { DaffModalComponent } from '../../modal/modal.component';
 import { DaffModalService } from '../modal.service';
 
 @Component({
+  selector: 'daff-dynamic-component',
   template: `
 		<p>It works!</p>
 	`,
@@ -23,6 +24,14 @@ import { DaffModalService } from '../modal.service';
 })
 class DynamicComponent {}
 
+@Component({
+  selector: 'daff-dynamic-component-two',
+  template: `
+		<p>It works!</p>
+	`,
+  standalone: true,
+})
+class DynamicTwoComponent {}
 
 @Component({
   template: '',
@@ -33,6 +42,7 @@ class WrapperComponent {
 }
 
 describe('@daffodil/design/modal | DaffModalService', () => {
+  let modalService: DaffModalService;
   let service: DaffModalService;
   let overlayContainer: OverlayContainer;
   let overlayContainerElement: HTMLElement;
@@ -43,10 +53,10 @@ describe('@daffodil/design/modal | DaffModalService', () => {
     TestBed.configureTestingModule({
       imports: [
         DynamicComponent,
+        DynamicTwoComponent,
         DaffModalComponent,
         NoopAnimationsModule,
         WrapperComponent,
-
       ],
       providers: [
         DaffModalService,
@@ -61,6 +71,7 @@ describe('@daffodil/design/modal | DaffModalService', () => {
     fixture = TestBed.createComponent(WrapperComponent);
     fixture.detectChanges();
     wrapper = fixture.componentInstance;
+    modalService = TestBed.inject(DaffModalService);
   });
 
   it('should be created', () => {
@@ -88,6 +99,30 @@ describe('@daffodil/design/modal | DaffModalService', () => {
       expect(count).toEqual(1);
       expect(overlayContainerElement.textContent).toContain('It works!');
     });
+
+    it('should close any previously opened modals', () => {
+      modalService.open(DynamicComponent);
+      modalService.open(DynamicComponent);
+
+      expect(
+        overlayContainerElement.querySelectorAll('daff-dynamic-component').length,
+      ).toEqual(1);
+    });
+
+    it('should close all modals except the last one opened', fakeAsync(() => {
+      modalService.open(DynamicComponent);
+      modalService.open(DynamicTwoComponent);
+
+      fixture.detectChanges();
+      flush();
+
+      expect(
+        overlayContainerElement.querySelector('daff-dynamic-component-two'),
+      ).toBeDefined();
+      expect(
+        overlayContainerElement.querySelector('daff-dynamic-component'),
+      ).toBeNull();
+    }));
   });
 
   describe('the default configuration', () => {
